+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create attribute service that allows to fetch attributes with specific options - used for products aggregates - @gibkigonzo (https://github.com/DivanteLtd/vue-storefront/pull/4001, https://github.com/DivanteLtd/mage2vuestorefront/pull/99)
 - Add ElasticSearch client support for HTTP authentication - @cewald (#397)
 - Endpoint for reset password with reset token. Only for Magento 2 - @Fifciu
+- Varnish Cache with autoinvalidation by Cache tags as addon - @Fifciu
 
 ### Fixed
 - add es7 support for map url module and fixed default index for es config - @gibkigonzo